+++ models/getAll.go
@@ -4,7 +4,7 @@ import (
 	"postgressql/db"
 )
 
-func GetAll(id int) (todos []Todo, err error){
+func GetAll(id int64) (todos []Todo, err error) {
 	conn, err := db.OpenConnection()
 	if err != nil {
 		return
@@ -15,7 +15,7 @@ func GetAll(id int) (todos []Todo, err error){
 	if err != nil {
 		return
 	}
-	for rows.Next(){
+	for rows.Next() {
 		var todo Todo
 		err = rows.Scan(&todo.ID, &todo.Title, &todo.Description, &todo.Done)
 		if err != nil {
